Move image into importer folder
